fixed warnings with third party libraries

+++ godEngine/OpenGL/Internal/OGLPrimitives.cpp
@@ -55,11 +55,11 @@ namespace god
 		glEnableVertexAttribArray ( 5 );
 		glVertexAttribPointer ( 5 , 4 , GL_FLOAT , GL_FALSE , 4 * vec4Size , ( void* ) 0 );
 		glEnableVertexAttribArray ( 6 );
-		glVertexAttribPointer ( 6 , 4 , GL_FLOAT , GL_FALSE , 4 * vec4Size , ( void* ) ( 1 * vec4Size ) );
+		glVertexAttribPointer ( 6 , 4 , GL_FLOAT , GL_FALSE , 4 * vec4Size , ( void* ) static_cast< long long >( 1 * vec4Size ) );
 		glEnableVertexAttribArray ( 7 );
-		glVertexAttribPointer ( 7 , 4 , GL_FLOAT , GL_FALSE , 4 * vec4Size , ( void* ) ( 2 * vec4Size ) );
+		glVertexAttribPointer ( 7 , 4 , GL_FLOAT , GL_FALSE , 4 * vec4Size , ( void* ) static_cast< long long >( 2 * vec4Size ) );
 		glEnableVertexAttribArray ( 8 );
-		glVertexAttribPointer ( 8 , 4 , GL_FLOAT , GL_FALSE , 4 * vec4Size , ( void* ) ( 3 * vec4Size ) );
+		glVertexAttribPointer ( 8 , 4 , GL_FLOAT , GL_FALSE , 4 * vec4Size , ( void* ) static_cast< long long >( 3 * vec4Size ) );
 		glVertexAttribDivisor ( 5 , 1 );
 		glVertexAttribDivisor ( 6 , 1 );
 		glVertexAttribDivisor ( 7 , 1 );
